try multi-canister testing
