+++ examples/fib/src/main.rs
@@ -1,0 +1,2 @@
+mod canister;
+fn main() {}
