+++ examples/multi_counter/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "ic_kit_example_fib"
+name = "ic_kit_example_multi_counter"
 version = "0.1.0"
 edition = "2021"
 
@@ -7,7 +7,8 @@ edition = "2021"
 
 [dependencies]
 ic-kit = {path="../../ic-kit"}
+ic_kit_example_counter = {path="../counter"}
 
 [[bin]]
-name = "ic_kit_example_fib"
+name = "ic_kit_example_multi_counter"
 path = "src/main.rs"
