+++ examples/multi_counter/candid.did
@@ -1,0 +1,1 @@
+service : { increment : () -> () }
