+++ examples/multi_counter/src/lib.rs
@@ -1,0 +1,5 @@
+//! A simple example of a counter service that can proxy an increment call to a list of counter
+//! canister.
+
+pub mod canister;
+pub use canister::MultiCounterCanister;
